parser/requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
spacy
pydantic
